requirements.txt
requests==2.24.0
python-twitter==3.5
htmldom==2.0
lxml
